+++ generators/app/templates/src/app.js
@@ -1,22 +1,23 @@
 /*
  * The `app` module
- * ============================================================================
+ * ================
  *
- * This module provides the game initialization routine.
+ * Provides the game initialization routine.
  */
 
 // Required: import the Babel runtime module.
 import 'babel-polyfill';
 
 // Import game states.
-import * as states from './app/states';
-
+import * as states from './states';
 
 export function init () {
   const game = new Phaser.Game(640, 480, Phaser.AUTO);
 
   // Dynamically add all required game states.
-  Object.keys(states).forEach((key) => game.state.add(key, states[key]));
+  Object
+    .entries(states)
+    .forEach(([key, state]) => game.state.add(key, state));
 
   game.state.start('Boot');